Refactor Routing #1
