+++ server/cronJob.js
@@ -1,9 +1,9 @@
 const cron = require('node-cron')
 const eventBus = require('./src/events/eventBus')
 const moment = require("moment");
-const { createOrUpdateTeams, updateTeamsRanking } = require('./src/controllers/teamController')
-const { updateMatchStatusAndPredictions } = require('./src/controllers/matchController')
-const { checkupBets } = require('./src/controllers/betController')
+const { createOrUpdateTeams, updateTeamsRanking } = require('./src/services/teamService')
+const { updateMatchStatusAndPredictions } = require('./src/services/matchService')
+const { checkupBets } = require('./src/services/betService')
 const { updatePlayers } = require('./src/controllers/playerController')
 const { checkAndScheduleSeasonEndTasks } = require("./src/services/appService");
 const { updateMatches, fetchWeekMatches } = require('./src/services/matchService');
