+++ server/src/controllers/playerController.js
@@ -1,112 +1,42 @@
-const {Team, Player, PlayerTeamCompetition} = require("../models");
-const axios = require("axios");
-const {getCurrentSeasonId, getCurrentSeasonYear} = require("./seasonController");
-const logger = require("../utils/logger/logger");
-const apiKey = process.env.FB_API_KEY;
-const apiHost = process.env.FB_API_HOST;
-const apiBaseUrl = process.env.FB_API_URL;
+const express = require('express')
+const router = express.Router()
+const {authenticateJWT} = require("../middlewares/auth")
+const {Team, Player, PlayerTeamCompetition} = require("../models")
+const axios = require("axios")
+const {getCurrentSeasonId, getCurrentSeasonYear} = require("../services/seasonService")
+const logger = require("../utils/logger/logger")
+const apiKey = process.env.FB_API_KEY
+const apiHost = process.env.FB_API_HOST
+const apiBaseUrl = process.env.FB_API_URL
 
-async function updatePlayers(teamIds = [], competitionId = null) {
+router.get('/players', authenticateJWT, async (req, res) => {
   try {
-    let teams = [];
-    if (!Array.isArray(teamIds)) {
-      teamIds = [teamIds];
-    }
-    if (teamIds.length > 0) {
-      teams = await Team.findAll({
-        where: { id: teamIds }
-      });
+    const teamId1 = req.query.teamId1;
+    const teamId2 = req.query.teamId2;
+    let queryCondition;
+    if (teamId1 && teamId2) {
+      queryCondition = {
+        teamId: [teamId1, teamId2]
+      };
+    } else if (teamId1 || teamId2) {
+      queryCondition = {
+        teamId: teamId1 || teamId2
+      };
     } else {
-      teams = await Team.findAll();
-    }
-    const seasonYear = await getCurrentSeasonYear(competitionId)
-    for (const team of teams) {
-      let currentPage = 1;
-      let totalPages = 0;
-      do {
-        const options = {
-          method: 'GET',
-          url: `${apiBaseUrl}players/`,
-          params: {
-            team: `${team.id}`,
-            season: seasonYear,
-            page: currentPage
-          },
-          headers: {
-            'X-RapidAPI-Key': apiKey,
-            'X-RapidAPI-Host': apiHost
-          }
-        };
-        const response = await axios.request(options);
-        const apiPlayers = response.data.response;
-        for (const apiPlayer of apiPlayers) {
-          let [player, created] = await Player.findOrCreate({
-            where: { id: apiPlayer.player.id },
-            defaults: {
-              name: apiPlayer.player.name,
-              firstname: apiPlayer.player.firstname,
-              lastname: apiPlayer.player.lastname,
-              photo: apiPlayer.player.photo
-            }
-          });
-          if (!created) {
-            await Player.update({
-              name: apiPlayer.player.name,
-              firstname: apiPlayer.player.firstname,
-              lastname: apiPlayer.player.lastname,
-              photo: apiPlayer.player.photo
-            }, {
-              where: { id: apiPlayer.player.id }
-            });
-          }
-          const association = await PlayerTeamCompetition.findOne({
-            where: {
-              playerId: apiPlayer.player.id,
-              teamId: team.id,
-              competitionId: competitionId
-            }
-          });
-          if (!association) {
-            await PlayerTeamCompetition.create({
-              playerId: apiPlayer.player.id,
-              teamId: team.id,
-              competitionId: competitionId
-            });
-          }
-        }
-
-        totalPages = response.data.paging.total;
-        currentPage++;
-      } while (currentPage <= totalPages);
-    }
-    logger.info(`Mise à jour des joueurs effectuées avec succès`);
-  } catch (error) {
-    logger.error(`Erreur lors de le récupération des joueurs: `, error);
-  }
-}
-const getPlayersByTeamId = async (teamId) => {
-  try {
-    if (!teamId) {
-      logger.info('teamId manquant');
-      return;
+      return res.status(400).send('Aucun identifiant d\'équipe fourni');
     }
     const players = await PlayerTeamCompetition.findAll({
-      where: {
-        teamId: teamId
-      },
-      include: {
-        model: Player,
-        as: 'Player',
-      }
+      where: queryCondition,
+      include: [
+        { model: Player, as: 'Player' },
+        { model: Team, as: 'Team' },
+      ]
     });
-    logger.info('Joueurs récupérés : ', players);
-    return players;
+    res.json(players);
   } catch (error) {
-    logger.error(`Erreur lors de le récupération des joueurs: `, error);
+    console.error('Erreur lors de la récupération des joueurs :', error);
+    res.status(500).send('Erreur interne du serveur');
   }
-}
+})
 
-module.exports = {
-  updatePlayers,
-  getPlayersByTeamId
-};
+module.exports = router
