+++ server/src/controllers/rewardController.js
@@ -1,29 +1,17 @@
+const express = require('express')
+const router = express.Router()
+const {authenticateJWT} = require("../middlewares/auth");
 const rewardService = require('../services/rewardService');
 const logger = require("../utils/logger/logger");
+const {Reward} = require("../models");
 
-const getAllRewards = async function (req, res) {
+router.get('/rewards', authenticateJWT, async (req, res) => {
   try {
-    const rewards = rewardService.allRewards();
+    const rewards = await Reward.findAll();
     res.json(rewards);
   } catch (error) {
     res.status(500).json({ message: 'Route protégée', error: error.message })
-    logger.error('Erreur lors de la recuperation des trophées:', error)
   }
-}
+})
 
-const getAllUserRewards = async function (req, res) {
-  const userId = req.params.userId;
-  const rewards = await rewardService.getRewardsByUser(userId);
-  res.json(rewards);
-}
-
-const checkAvailableRewards = async function (req, res) {
-  const userId = req.params.userId;
-  const rewards = await rewardService.checkAvailableRewards(userId);
-  res.json(rewards);
-}
-
-module.exports = {
-  getAllRewards,
-  getAllUserRewards,
-}
+module.exports = router
