+++ server/src/middlewares/auth.js
@@ -1,0 +1,30 @@
+const jwt = require("jsonwebtoken");
+const secretKey = process.env.SECRET_KEY
+
+const authenticateJWT = (req, res, next) => {
+  const authHeader = req.headers.authorization;
+
+  if (!authHeader) {
+    return res.status(401).json({ error: 'Authorization header is missing' });
+  }
+
+  const token = authHeader.split(' ')[1]; // Extract the token from Bearer
+  if (!token) {
+    return res.status(401).json({ error: 'Token is missing from Authorization header' });
+  }
+
+  jwt.verify(token, secretKey, (err, user) => {
+    if (err) {
+      let errorMsg = 'Invalid token';
+      if (err.name === 'TokenExpiredError') errorMsg = 'Token has expired';
+      return res.status(403).json({ error: errorMsg });
+    }
+
+    req.user = user;
+    next();
+  });
+};
+
+module.exports = {
+  authenticateJWT
+};
