+++ server/src/services/appService.js
@@ -4,11 +4,36 @@ const apiHost = process.env.FB_API_HOST;
 const apiBaseUrl = process.env.FB_API_URL;
 const logger = require("../utils/logger/logger");
 const {Op} = require("sequelize");
-const {Season} = require("../models");
+const {Season, Setting} = require("../models");
 const schedule = require('node-schedule');
 const {checkSeasonRewards} = require("./rewardService");
+const moment = require("moment/moment");
 
-const getCurrentMatchday = async function () {
+const getAPICallsCount = async () => {
+  try {
+    const options = {
+      method: 'GET',
+      url: `${apiBaseUrl}status/`,
+      headers: {
+        'X-RapidAPI-Key': apiKey,
+        'X-RapidAPI-Host': apiHost
+      }
+    };
+    const response = await axios.request(options);
+    return response.data;
+  } catch (error) {
+    console.log('Erreur lors de la récupération des appels API : ', error);
+  }
+}
+
+const getMonthDateRange = () => {
+  const moment = require('moment');
+  const start = moment().startOf('month');
+  const end = moment().endOf('month');
+  return { start: start, end: end };
+}
+
+const getCurrentMatchday = async () => {
   try {
     const response = await Season.findAll({
       where: {
@@ -53,8 +78,14 @@ const checkAndScheduleSeasonEndTasks = async () => {
   }
 };
 
+const getSettlement = async () => {
+  return await Setting.findOne({where: {key: 'regulation'}})
+}
 
 module.exports = {
+  getAPICallsCount,
+  getMonthDateRange,
   getCurrentMatchday,
   checkAndScheduleSeasonEndTasks,
+  getSettlement,
 }
