+++ server/src/services/matchService.js
@@ -3,13 +3,48 @@ const apiKey = process.env.FB_API_KEY;
 const apiHost = process.env.FB_API_HOST;
 const apiBaseUrl = process.env.FB_API_URL;
 const {Match} = require("../models");
-const {getCurrentSeasonId, getCurrentSeasonYear} = require("../controllers/seasonController");
+const {getCurrentSeasonId, getCurrentSeasonYear} = require("../services/seasonService");
 const ProgressBar = require("progress");
 const {Op} = require("sequelize");
 const {schedule} = require("node-schedule");
-const {updateMatchStatusAndPredictions} = require("../controllers/matchController");
 const {checkBetByMatchId} = require("./betService");
 const moment = require("moment");
+const {getMonthDateRange} = require("../controllers/appController");
+let cronTasks = [];
+
+async function updateMatchStatusAndPredictions(matchIds) {
+  if (!Array.isArray(matchIds)) {
+    matchIds = [matchIds];
+  }
+  for (const matchId of matchIds) {
+    await updateSingleMatch(matchId);
+  }
+}
+
+async function getCurrentMonthMatchdays() {
+  try {
+    const matchdays = []
+    const monthDates = getMonthDateRange();
+    const matchs = await Match.findAll({
+      where: {
+        utcDate: {
+          [Op.gte]: monthDates.start,
+          [Op.lte]: monthDates.end
+        }
+      }
+    })
+    for (const match of matchs) {
+      matchdays.push(match.matchday)
+    }
+    return matchdays
+  } catch (error) {
+    console.log( 'Erreur lors de la récupération des matchs du mois courant:', error)
+  }
+}
+
+function getMatchsCronTasks() {
+  return cronTasks
+}
 
 async function updateSingleMatch(matchId) {
   try {
@@ -190,6 +225,7 @@ async function fetchWeekMatches() {
 }
 
 module.exports = {
+  updateMatchStatusAndPredictions,
   updateSingleMatch,
   updateMatches,
   fetchWeekMatches
